MODBULKOPS-131 - administrative and items notes updater
