+++ src/main/java/org/folio/bulkops/processor/HoldingsNotesUpdater.java
@@ -1,10 +1,12 @@
 package org.folio.bulkops.processor;
 
+import lombok.AllArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
 import org.folio.bulkops.domain.bean.HoldingsNote;
 import org.folio.bulkops.domain.bean.HoldingsRecord;
 import org.folio.bulkops.domain.dto.Action;
 import org.folio.bulkops.domain.dto.Parameter;
+import org.folio.bulkops.domain.dto.UpdateOptionType;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
@@ -12,13 +14,62 @@ import java.util.List;
 import java.util.Optional;
 
 import static java.util.stream.Collectors.toCollection;
+import static org.folio.bulkops.domain.dto.UpdateActionType.ADD_TO_EXISTING;
+import static org.folio.bulkops.domain.dto.UpdateActionType.CHANGE_TYPE;
+import static org.folio.bulkops.domain.dto.UpdateActionType.FIND_AND_REMOVE_THESE;
+import static org.folio.bulkops.domain.dto.UpdateActionType.FIND_AND_REPLACE;
+import static org.folio.bulkops.domain.dto.UpdateActionType.MARK_AS_STAFF_ONLY;
+import static org.folio.bulkops.domain.dto.UpdateActionType.REMOVE_ALL;
+import static org.folio.bulkops.domain.dto.UpdateActionType.REMOVE_MARK_AS_STAFF_ONLY;
 import static org.folio.bulkops.domain.dto.UpdateOptionType.ADMINISTRATIVE_NOTE;
+import static org.folio.bulkops.domain.dto.UpdateOptionType.HOLDINGS_NOTE;
 
 @Component
+@AllArgsConstructor
 public class HoldingsNotesUpdater {
-  public static final String HOLDINGS_NOTE_TYPE_ID_KEY = "HOLDINGS_NOTE_TYPE_ID_KEY";
 
-  public void setMarkAsStaffForNotesByTypeId(List<HoldingsNote> notes, List<Parameter> parameters, boolean markAsStaffValue) {
+  public static final String HOLDINGS_NOTE_TYPE_ID_KEY = "HOLDINGS_NOTE_TYPE_ID_KEY";
+  private final AdministrativeNotesUpdater administrativeNotesUpdater;
+
+  public Optional<Updater<HoldingsRecord>> updateNotes( Action action, UpdateOptionType option) {
+    if ((MARK_AS_STAFF_ONLY == action.getType() || REMOVE_MARK_AS_STAFF_ONLY == action.getType()) && option == HOLDINGS_NOTE){
+      var markAsStaffValue = action.getType() == MARK_AS_STAFF_ONLY;
+      return Optional.of(holding -> setMarkAsStaffForNotesByTypeId(holding.getNotes(), action.getParameters(), markAsStaffValue));
+    } else if (REMOVE_ALL == action.getType()) {
+      if (option == ADMINISTRATIVE_NOTE) {
+        return Optional.of(holding -> holding.setAdministrativeNotes(administrativeNotesUpdater.removeAdministrativeNotes()));
+      } else if (option == HOLDINGS_NOTE) {
+        return Optional.of(holding -> holding.setNotes(removeNotesByTypeId(holding.getNotes(), action.getParameters())));
+      }
+    } else if (ADD_TO_EXISTING == action.getType()) {
+      if (option == ADMINISTRATIVE_NOTE) {
+        return Optional.of(holding -> holding.setAdministrativeNotes(administrativeNotesUpdater.addToAdministrativeNotes(action.getUpdated(), holding.getAdministrativeNotes())));
+      } else if (option == HOLDINGS_NOTE) {
+        return Optional.of(holding -> holding.setNotes(addToNotesByTypeId(holding.getNotes(), action.getParameters(), action.getUpdated())));
+      }
+    } else if (FIND_AND_REMOVE_THESE == action.getType()) {
+      if (option == ADMINISTRATIVE_NOTE) {
+        return Optional.of(holding -> holding.setAdministrativeNotes(administrativeNotesUpdater.findAndRemoveAdministrativeNote(action.getInitial(), holding.getAdministrativeNotes())));
+      } else if (option == HOLDINGS_NOTE) {
+        return Optional.of(holding -> holding.setNotes(findAndRemoveNoteByValueAndTypeId(action.getInitial(), holding.getNotes(), action.getParameters())));
+      }
+    } else if (FIND_AND_REPLACE == action.getType()) {
+      if (option == ADMINISTRATIVE_NOTE) {
+        return Optional.of(holding -> holding.setAdministrativeNotes(administrativeNotesUpdater.findAndReplaceAdministrativeNote(action, holding.getAdministrativeNotes())));
+      } else if (option == HOLDINGS_NOTE) {
+        return Optional.of(holding -> findAndReplaceNoteByValueAndTypeId(action, holding.getNotes()));
+      }
+    } else if (CHANGE_TYPE == action.getType()) {
+      if (option == ADMINISTRATIVE_NOTE) {
+        return Optional.of(holding -> administrativeNotesUpdater.changeNoteTypeForAdministrativeNotes(holding, action));
+      } else if (option == HOLDINGS_NOTE) {
+        return Optional.of(holding -> changeNoteTypeForHoldingsNote(holding, action));
+      }
+    }
+    return Optional.empty();
+  }
+
+  private void setMarkAsStaffForNotesByTypeId(List<HoldingsNote> notes, List<Parameter> parameters, boolean markAsStaffValue) {
     parameters.stream().filter(p -> StringUtils.equals(p.getKey(), HOLDINGS_NOTE_TYPE_ID_KEY)).findFirst()
       .ifPresent(parameter -> {
         var typeId = parameter.getValue();
@@ -31,11 +82,7 @@ public class HoldingsNotesUpdater {
         }});
   }
 
-  public List<String> removeAdministrativeNotes() {
-    return new ArrayList<>();
-  }
-
-  public List<HoldingsNote> removeNotesByTypeId(List<HoldingsNote> notes, List<Parameter> parameters) {
+  private List<HoldingsNote> removeNotesByTypeId(List<HoldingsNote> notes, List<Parameter> parameters) {
     var typeIdParameterOptional = getTypeIdParameterOptional(parameters);
     if (typeIdParameterOptional.isPresent()) {
       var typeId = typeIdParameterOptional.get().getValue();
@@ -46,15 +93,7 @@ public class HoldingsNotesUpdater {
     return notes;
   }
 
-  public List<String> addToAdministrativeNotes(String administrativeNote, List<String> administrativeNotes) {
-    if (administrativeNotes == null) {
-      administrativeNotes = new ArrayList<>();
-    }
-    administrativeNotes.add(administrativeNote);
-    return administrativeNotes;
-  }
-
-  public List<HoldingsNote> addToNotesByTypeId(List<HoldingsNote> notes, List<Parameter> parameters, String noteValue) {
+  private List<HoldingsNote> addToNotesByTypeId(List<HoldingsNote> notes, List<Parameter> parameters, String noteValue) {
     var typeIdParameterOptional = getTypeIdParameterOptional(parameters);
     if (typeIdParameterOptional.isPresent()) {
       var note = new HoldingsNote().withHoldingsNoteTypeId(typeIdParameterOptional.get().getValue())
@@ -67,16 +106,7 @@ public class HoldingsNotesUpdater {
     return notes;
   }
 
-  public List<String> findAndRemoveAdministrativeNote(String administrativeNoteToRemove, List<String> administrativeNotes) {
-    if (administrativeNotes != null) {
-      administrativeNotes = administrativeNotes.stream()
-        .filter(administrativeNote -> !StringUtils.equals(administrativeNote, administrativeNoteToRemove))
-        .collect(toCollection(ArrayList::new));
-    }
-    return administrativeNotes;
-  }
-
-  public List<HoldingsNote> findAndRemoveNoteByValueAndTypeId(String noteToRemove, List<HoldingsNote> notes, List<Parameter> parameters) {
+  private List<HoldingsNote> findAndRemoveNoteByValueAndTypeId(String noteToRemove, List<HoldingsNote> notes, List<Parameter> parameters) {
     var typeIdParameterOptional = getTypeIdParameterOptional(parameters);
     if (typeIdParameterOptional.isPresent() && notes != null) {
       notes = notes.stream().filter(note -> !(StringUtils.equals(note.getHoldingsNoteTypeId(), typeIdParameterOptional.get().getValue())
@@ -85,19 +115,7 @@ public class HoldingsNotesUpdater {
     return notes;
   }
 
-  public List<String> findAndReplaceAdministrativeNote(Action action, List<String> administrativeNotes) {
-    if (administrativeNotes != null) {
-      administrativeNotes = administrativeNotes.stream().map(administrativeNote -> {
-        if (StringUtils.equals(administrativeNote, action.getInitial())) {
-          return action.getUpdated();
-        }
-        return administrativeNote;
-      }).collect(toCollection(ArrayList::new));
-    }
-    return administrativeNotes;
-  }
-
-  public void findAndReplaceNoteByValueAndTypeId(Action action, List<HoldingsNote> notes) {
+  private void findAndReplaceNoteByValueAndTypeId(Action action, List<HoldingsNote> notes) {
     var typeIdParameterOptional = getTypeIdParameterOptional(action.getParameters());
     if (typeIdParameterOptional.isPresent() && notes != null) {
       notes.forEach(note -> {
@@ -107,16 +125,7 @@ public class HoldingsNotesUpdater {
     }
   }
 
-  public void changeNoteTypeForAdministrativeNotes(HoldingsRecord holding, Action action) {
-    if (holding.getAdministrativeNotes() != null) {
-      if (holding.getNotes() == null) holding.setNotes(new ArrayList<>());
-      holding.getAdministrativeNotes().forEach(administrativeNote ->
-        holding.getNotes().add(new HoldingsNote().withHoldingsNoteTypeId(action.getUpdated()).withNote(administrativeNote).withStaffOnly(false)));
-      holding.setAdministrativeNotes(new ArrayList<>());
-    }
-  }
-
-  public void changeNoteTypeForHoldingsNote(HoldingsRecord holding, Action action) {
+  private void changeNoteTypeForHoldingsNote(HoldingsRecord holding, Action action) {
     var typeIdParameterOptional = getTypeIdParameterOptional(action.getParameters());
     if (typeIdParameterOptional.isPresent()) {
       var typeId  = typeIdParameterOptional.get().getValue();
